Update to integration

+++ CHAPSim_post/_libs/integrate.pyx
@@ -147,7 +147,7 @@ def IntegrateTrapz(np.ndarray input_array,np.ndarray x, int axis=0, bint stagger
     cdef int i, j=0
 
     if dim == 1:
-        shape = np.ones(1)
+        shape = np.ones(1,dtype=int)
     else:
         shape = np.zeros(dim-1,dtype=int)
         for i in range(dim):
